Add files via upload
the .py will ask for the name of the csv file so enter it with the extension. The cntfile is basically a global counter to make sure that the insert into the db doesnt overwrite the data as we seem to have multiple csv files in the database. Change the value of cntfile to 0 when u start.